fix: min_active_bid returned as NumberOrHex
