+++ state-chain/custom-rpc/src/lib.rs
@@ -1,4 +1,4 @@
-use crate::boost_pool_rpc::BoostPoolFeesRpc;
+use crate::{boost_pool_rpc::BoostPoolFeesRpc, monitoring::RpcEpochStateV2};
 use boost_pool_rpc::BoostPoolDetailsRpc;
 use cf_amm::{
 	common::{Amount as AmmAmount, PoolPairsMap, Side, Tick},
@@ -46,7 +46,7 @@ use state_chain_runtime::{
 	chainflip::{BlockUpdate, Offence},
 	constants::common::TX_FEE_MULTIPLIER,
 	monitoring_apis::{
-		ActivateKeysBroadcastIds, AuthoritiesInfo, BtcUtxos, EpochState, ExternalChainsBlockHeight,
+		ActivateKeysBroadcastIds, AuthoritiesInfo, BtcUtxos, ExternalChainsBlockHeight,
 		FeeImbalance, FlipSupply, LastRuntimeUpgradeInfo, MonitoringData, OpenDepositChannels,
 		PendingBroadcasts, PendingTssCeremonies, RedemptionsInfo, SolanaNonces,
 	},
@@ -67,26 +67,7 @@ use std::{
 pub mod monitoring;
 pub mod order_fills;
 
-#[derive(Serialize, Deserialize, Clone)]
-pub struct RpcEpochState {
-	pub epoch_duration: u32,
-	pub current_epoch_started_at: u32,
-	pub current_epoch_index: u32,
-	pub min_active_bid: Option<NumberOrHex>,
-	pub rotation_phase: String,
-}
-impl From<EpochState> for RpcEpochState {
-	fn from(rotation_state: EpochState) -> Self {
-		Self {
-			epoch_duration: rotation_state.epoch_duration,
-			current_epoch_started_at: rotation_state.current_epoch_started_at,
-			current_epoch_index: rotation_state.current_epoch_index,
-			rotation_phase: rotation_state.rotation_phase,
-			min_active_bid: rotation_state.min_active_bid.map(Into::into),
-		}
-	}
-}
-#[derive(Serialize, Deserialize, Clone)]
+#[derive(Serialize, Deserialize)]
 pub struct RpcRedemptionsInfo {
 	pub total_balance: NumberOrHex,
 	pub count: u32,
@@ -117,7 +98,7 @@ impl From<FlipSupply> for RpcFlipSupply {
 pub struct RpcMonitoringData {
 	pub external_chains_height: ExternalChainsBlockHeight,
 	pub btc_utxos: BtcUtxos,
-	pub epoch: RpcEpochState,
+	pub epoch: RpcEpochStateV2,
 	pub pending_redemptions: RpcRedemptionsInfo,
 	pub pending_broadcasts: PendingBroadcasts,
 	pub pending_tss: PendingTssCeremonies,
