+++ state-chain/custom-rpc/src/monitoring.rs
@@ -1,6 +1,7 @@
 use super::pass_through;
 use crate::{BlockT, CustomRpc, RpcAccountInfoV2, RpcFeeImbalance, RpcMonitoringData, RpcResult};
 use cf_chains::{dot::PolkadotAccountId, sol::SolAddress};
+use cf_utilities::rpc::NumberOrHex;
 use jsonrpsee::proc_macros::rpc;
 use sc_client_api::{BlockchainEvents, HeaderBackend};
 use serde::{Deserialize, Serialize};
@@ -13,23 +14,42 @@ use state_chain_runtime::{
 		PendingTssCeremonies, RedemptionsInfo, SolanaNonces,
 	},
 };
+impl From<EpochState> for RpcEpochState {
+	fn from(rotation_state: EpochState) -> Self {
+		Self {
+			epoch_duration: rotation_state.epoch_duration,
+			current_epoch_started_at: rotation_state.current_epoch_started_at,
+			current_epoch_index: rotation_state.current_epoch_index,
+			rotation_phase: rotation_state.rotation_phase,
+			min_active_bid: rotation_state.min_active_bid.map(Into::into),
+		}
+	}
+}
+#[derive(Serialize, Deserialize)]
+pub struct RpcEpochState {
+	pub epoch_duration: u32,
+	pub current_epoch_started_at: u32,
+	pub current_epoch_index: u32,
+	pub min_active_bid: Option<NumberOrHex>,
+	pub rotation_phase: String,
+}
 
 // Temporary struct to hold the deprecated blocks_per_epoch field.
 // Can be deleted after v1.7 is released (meaning: after the version is bumped to 1.8).
 #[derive(Serialize, Deserialize)]
-pub struct RpcEpochState {
+pub struct RpcEpochStateV2 {
 	#[deprecated(
 		since = "1.8.0",
 		note = "This field is deprecated and will be removed in v1.8. Use blocks_per_epoch instead."
 	)]
 	blocks_per_epoch: u32,
 	#[serde(flatten)]
-	epoch_state: EpochState,
+	epoch_state: RpcEpochState,
 }
 
-impl From<EpochState> for RpcEpochState {
+impl From<EpochState> for RpcEpochStateV2 {
 	fn from(epoch_state: EpochState) -> Self {
-		Self { blocks_per_epoch: epoch_state.epoch_duration, epoch_state }
+		Self { blocks_per_epoch: epoch_state.epoch_duration, epoch_state: epoch_state.into() }
 	}
 }
 
@@ -52,7 +72,7 @@ pub trait MonitoringApi {
 		at: Option<state_chain_runtime::Hash>,
 	) -> RpcResult<Vec<(Offence, u32)>>;
 	#[method(name = "epoch_state")]
-	fn cf_epoch_state(&self, at: Option<state_chain_runtime::Hash>) -> RpcResult<RpcEpochState>;
+	fn cf_epoch_state(&self, at: Option<state_chain_runtime::Hash>) -> RpcResult<RpcEpochStateV2>;
 	#[method(name = "redemptions")]
 	fn cf_redemptions(&self, at: Option<state_chain_runtime::Hash>) -> RpcResult<RedemptionsInfo>;
 	#[method(name = "pending_broadcasts")]
@@ -121,7 +141,7 @@ where
 		cf_btc_utxos() -> BtcUtxos,
 		cf_dot_aggkey() -> PolkadotAccountId,
 		cf_suspended_validators() -> Vec<(Offence, u32)>,
-		cf_epoch_state() -> RpcEpochState [map: RpcEpochState::from],
+		cf_epoch_state() -> RpcEpochStateV2 [map: RpcEpochStateV2::from],
 		cf_redemptions() -> RedemptionsInfo,
 		cf_pending_broadcasts_count() -> PendingBroadcasts,
 		cf_pending_tss_ceremonies_count() -> PendingTssCeremonies,
